+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Various intrinsic routines had incorrect signatures around dynamic and open arrays.
 - False positives around platform-dependent binary expressions in `PlatformDependentTruncation`.
+- Incorrect type resolution around array property accesses.
 
 ## [1.12.0] - 2024-12-02